+++ test-utils-event-store/test-utils-persistence/src/test/java/uk/gov/justice/services/test/utils/persistence/TestJdbcDataSourceProviderTest.java
@@ -2,11 +2,8 @@ package uk.gov.justice.services.test.utils.persistence;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
-import static org.junit.Assert.fail;
 
 import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
 
 import javax.sql.DataSource;